requirements.txt
aiogram==3.20.0.post0
SQLAlchemy==2.0.40
aiomysql==0.2.0
python-dotenv==1.0.1
cryptography==44.0.2
PyMySQL==1.1.1
